Reorganize source code
There was a desire from the team to make this a multi-module
repository so HCP components can be released independently.

Thankfully the recent Go Workspaces feature makes this feasible
without constant go.mod modifications (see ./go.work).

There are now two Go modules:

github.com/Azure/ARO-HCP/frontend -> ./frontend
   The HTTP frontend component

github.com/Azure/ARO-HCP/internal -> ./internal
   Internal code that is or could potentially be shared across
   multiple releasable components like API definitions, utility
   packages, etc.
