+++ frontend/apiversions.go
@@ -6,5 +6,5 @@ package main
 // This will invoke the init() function in each
 // API version package so it can register itself.
 import (
-	_ "github.com/Azure/ARO-HCP/pkg/api/v20240610preview"
+	_ "github.com/Azure/ARO-HCP/internal/api/v20240610preview"
 )
