+++ frontend/frontend.go
@@ -12,7 +12,7 @@ import (
 	"os"
 	"sync/atomic"
 
-	"github.com/Azure/ARO-HCP/pkg/api/arm"
+	"github.com/Azure/ARO-HCP/internal/api/arm"
 )
 
 const (
