+++ frontend/go.mod
@@ -1,0 +1,3 @@
+module github.com/Azure/ARO-HCP/frontend
+
+go 1.22
