+++ frontend/middleware_validateapi.go
@@ -6,8 +6,8 @@ package main
 import (
 	"net/http"
 
-	"github.com/Azure/ARO-HCP/pkg/api"
-	"github.com/Azure/ARO-HCP/pkg/api/arm"
+	"github.com/Azure/ARO-HCP/internal/api"
+	"github.com/Azure/ARO-HCP/internal/api/arm"
 )
 
 func MiddlewareValidateAPIVersion(w http.ResponseWriter, r *http.Request, next http.HandlerFunc) {
