+++ internal/api/arm/resource_test.go
