+++ internal/api/hcpopenshiftcluster.go
@@ -6,8 +6,8 @@ package api
 import (
 	"net"
 
-	"github.com/Azure/ARO-HCP/pkg/api/arm"
-	"github.com/Azure/ARO-HCP/pkg/api/json"
+	"github.com/Azure/ARO-HCP/internal/api/arm"
+	"github.com/Azure/ARO-HCP/internal/api/json"
 )
 
 // HCPOpenShiftCluster represents an ARO HCP OpenShift cluster resource.
