+++ internal/api/v20240610preview/enums.go
