+++ internal/api/v20240610preview/enums_test.go
