+++ internal/api/v20240610preview/hcpopenshiftclusternodepool.go
@@ -4,8 +4,8 @@ package v20240610preview
 // Licensed under the Apache License 2.0.
 
 import (
-	"github.com/Azure/ARO-HCP/pkg/api"
-	"github.com/Azure/ARO-HCP/pkg/api/arm"
+	"github.com/Azure/ARO-HCP/internal/api"
+	"github.com/Azure/ARO-HCP/internal/api/arm"
 )
 
 // HCPOpenShiftClusterNodePool represents a node pool resource for ARO HCP
