+++ internal/api/v20240610preview/nodepoolprofile.go
