--- pkg/util/version/const.go
@@ -1,6 +1,0 @@
-package version
-
-// Copyright (c) Microsoft Corporation.
-// Licensed under the Apache License 2.0.
-
-var Version = "unknown" // overridden by Makefile
